In progress for EA-3, examples update
