+++ entity-services-examples/gen/ml-config/databases/content-database.json
@@ -1,1 +1,49 @@
-{"database-name":"%%DATABASE%%", "schema-database":"%%SCHEMAS_DATABASE%%", "path-namespace":[{"prefix":"es", "namespace-uri":"http://marklogic.com/entity-services"}], "element-word-lexicon":[{"collation":"http://marklogic.com/collation/en", "localname":"name", "namespace-uri":""}], "range-path-index":[{"collation":"http://marklogic.com/collation/en", "invalid-values":"reject", "path-expression":"//es:instance/Run/date", "range-value-positions":false, "scalar-type":"date"}, {"collation":"http://marklogic.com/collation/en", "invalid-values":"reject", "path-expression":"//es:instance/Run/distance", "range-value-positions":false, "scalar-type":"decimal"}, {"collation":"http://marklogic.com/collation/en", "invalid-values":"reject", "path-expression":"//es:instance/Run/duration", "range-value-positions":false, "scalar-type":"dayTimeDuration"}, {"collation":"http://marklogic.com/collation/en", "invalid-values":"reject", "path-expression":"//es:instance/Run/runByRunner", "range-value-positions":false, "scalar-type":"string"}], "triple-index":true, "collection-lexicon":true}
+{
+    "collection-lexicon": true,
+    "database-name": "%%DATABASE%%",
+    "element-word-lexicon": [
+        {
+            "collation": "http://marklogic.com/collation/en",
+            "localname": "name",
+            "namespace-uri": ""
+        }
+    ],
+    "path-namespace": [
+        {
+            "namespace-uri": "http://marklogic.com/entity-services",
+            "prefix": "es"
+        }
+    ],
+    "range-path-index": [
+        {
+            "collation": "http://marklogic.com/collation/en",
+            "invalid-values": "reject",
+            "path-expression": "//es:instance/Run/date",
+            "range-value-positions": false,
+            "scalar-type": "date"
+        },
+        {
+            "collation": "http://marklogic.com/collation/en",
+            "invalid-values": "reject",
+            "path-expression": "//es:instance/Run/distance",
+            "range-value-positions": false,
+            "scalar-type": "decimal"
+        },
+        {
+            "collation": "http://marklogic.com/collation/en",
+            "invalid-values": "reject",
+            "path-expression": "//es:instance/Run/duration",
+            "range-value-positions": false,
+            "scalar-type": "dayTimeDuration"
+        },
+        {
+            "collation": "http://marklogic.com/collation/en",
+            "invalid-values": "reject",
+            "path-expression": "//es:instance/Run/runByRunner",
+            "range-value-positions": false,
+            "scalar-type": "string"
+        }
+    ],
+    "schema-database": "%%SCHEMAS_DATABASE%%",
+    "triple-index": true
+}
