--- entity-services-examples/gen/ml-schemas/Race-0.0.1.tdex-gen
@@ -1,132 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<template xmlns="http://marklogic.com/xdmp/tde">
-  <description>
-            Extraction Template Generated from Entity Type Document
-            graph uri: http://grechaw.github.io/entity-types#Race-0.0.1</description>
-  <context>//es:instance</context>
-  <path-namespaces>
-    <path-namespace>
-      <prefix>es</prefix>
-      <namespace-uri>http://marklogic.com/entity-services</namespace-uri>
-    </path-namespace>
-  </path-namespaces>
-  <templates>
-    <template>
-      <context>./Race</context>
-      <rows xmlns:tde="http://marklogic.com/xdmp/tde">
-	<row>
-	  <schema-name>Race</schema-name>
-	  <view-name>Race</view-name>
-	  <columns>
-	    <column>
-	      <name>name</name>
-	      <scalar-type>string</scalar-type>
-	      <val>name</val>
-	    </column>
-	    <column>
-	      <name>wonByRunner</name>
-	      <scalar-type>string</scalar-type>
-	      <val>wonByRunner</val>
-	    </column>
-	    <column>
-	      <name>courseLength</name>
-	      <scalar-type>decimal</scalar-type>
-	      <val>courseLength</val>
-	    </column>
-	  </columns>
-	</row>
-      </rows>
-      <templates>
-	<template xmlns:tde="http://marklogic.com/xdmp/tde">
-	  <context>./comprisedOfRuns</context>
-	  <rows>
-	    <row>
-	      <schema-name>Race</schema-name>
-	      <view-name>Race_comprisedOfRuns</view-name>
-	      <columns>
-		<!-- this column joins to PK of 'parent' -->
-		<column>
-		  <name>name</name>
-		  <scalar-type>string</scalar-type>
-		  <val>../name</val>
-		</column>
-		<column>
-		  <name>comprisedOfRuns</name>
-		  <scalar-type>string</scalar-type>
-		  <val>.</val>
-		</column>
-	      </columns>
-	    </row>
-	  </rows>
-	</template>
-      </templates>
-    </template>
-    <template>
-      <context>./Run</context>
-      <rows xmlns:tde="http://marklogic.com/xdmp/tde">
-	<row>
-	  <schema-name>Race</schema-name>
-	  <view-name>Run</view-name>
-	  <columns>
-	    <column>
-	      <name>id</name>
-	      <scalar-type>string</scalar-type>
-	      <val>id</val>
-	    </column>
-	    <column>
-	      <name>date</name>
-	      <scalar-type>date</scalar-type>
-	      <val>date</val>
-	    </column>
-	    <column>
-	      <name>distance</name>
-	      <scalar-type>decimal</scalar-type>
-	      <val>distance</val>
-	    </column>
-	    <column>
-	      <name>distanceLabel</name>
-	      <scalar-type>string</scalar-type>
-	      <val>distanceLabel</val>
-	    </column>
-	    <column>
-	      <name>duration</name>
-	      <scalar-type>dayTimeDuration</scalar-type>
-	      <val>duration</val>
-	    </column>
-	    <column>
-	      <name>runByRunner</name>
-	      <scalar-type>string</scalar-type>
-	      <val>runByRunner</val>
-	    </column>
-	  </columns>
-	</row>
-      </rows>
-    </template>
-    <template>
-      <context>./Runner</context>
-      <rows xmlns:tde="http://marklogic.com/xdmp/tde">
-	<row>
-	  <schema-name>Race</schema-name>
-	  <view-name>Runner</view-name>
-	  <columns>
-	    <column>
-	      <name>name</name>
-	      <scalar-type>string</scalar-type>
-	      <val>name</val>
-	    </column>
-	    <column>
-	      <name>age</name>
-	      <scalar-type>int</scalar-type>
-	      <val>age</val>
-	    </column>
-	    <column>
-	      <name>gender</name>
-	      <scalar-type>string</scalar-type>
-	      <val>gender</val>
-	    </column>
-	  </columns>
-	</row>
-      </rows>
-    </template>
-  </templates>
-</template>
